MainViewModel: use BigInteger
in the worst case scenario, the value of WordValue can exceed te 64bit Long Integer limit, fix this by using BigInteger.

As it is not a primitive type, BigInteger is at the very least 15 times slower than Long and compounded by the fact that it uses lists for an infinitely big number, it makes the app much slower. I will consider doing parallelism later on.