MainViewModel: Fallback to sorted list comparison if WordValue exceeds integer limit.
BigInteger really really hurts performance as it uses lists. I reverted its use and have now implemented a hybrid solution. If the value hits overflow, change it to -1. If the value is -1 in GetAnagrams, this means that it is a value that has hit overflow and is impossible to process it mathematically. Hence we will fallback to the slower sorted string comparison method.